# backend/app/admin/service/cmd/server/assets/openapi.yaml
...
openapi: 3.0.3
info:
    title: Kratos巨石应用实践API
    description: Kratos巨石应用实践API
    contact:
        name: tx7do
        url: https://github.com/tx7do/kratos-monolithic-demo
        email: yanglinbo@gmail.com
    license:
        name: MIT License
        url: https://github.com/tx7do/kratos-monolithic-demo/blob/master/LICENSE
    version: "1.0"
paths:
...
    /admin/v1/login:
        post:
            tags:
                - AuthenticationService
            description: 登陆
            operationId: AuthenticationService_Login
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/LoginRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/LoginResponse'
    /admin/v1/logout:
        post:
            tags:
                - AuthenticationService
            description: 登出
            operationId: AuthenticationService_Logout
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/LogoutRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content: {}
            security:
                - OAuth2PasswordBearer: []
    /admin/v1/me:
        get:
            tags:
                - AuthenticationService
            description: 后台获取已经登陆的用户的数据
            operationId: AuthenticationService_GetMe
            parameters:
                - name: id
                  in: query
                  schema:
                    type: integer
                    format: uint32
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/User'
            security:
                - OAuth2PasswordBearer: []
...
    /admin/v1/refresh_token:
        post:
            tags:
                - AuthenticationService
            description: 后台登陆
            operationId: AuthenticationService_RefreshToken
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/RefreshTokenRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/LoginResponse'
...
components:
    schemas:
...
        LoginRequest:
            required:
                - username
                - password
                - grand_type
            type: object
            properties:
                username:
                    type: string
                password:
                    type: string
                grand_type:
                    type: string
                    default: password
                    description: 授权类型，此处的值固定为"password"，必选项。
                scope:
                    type: string
            description: 用户后台登陆 - 请求
        LoginResponse:
            type: object
            properties:
                access_token:
                    type: string
                refresh_token:
                    type: string
                token_type:
                    type: string
                expires_in:
                    type: string
            description: 用户后台登陆 - 回应
        LogoutRequest:
            type: object
            properties:
                id:
                    type: integer
                    format: uint32
            description: 用户后台登出 - 请求
...
        RefreshTokenRequest:
            required:
                - refresh_token
                - grand_type
            type: object
            properties:
                refresh_token:
                    type: string
                grand_type:
                    type: string
                    default: password
                    description: 授权类型，此处的值固定为"password"，必选项。
                scope:
                    type: string
            description: 用户刷新令牌 - 请求
...
        User:
            type: object
            properties:
                id:
                    type: integer
                    format: uint32
                roleId:
                    type: integer
                    description: 角色ID
                    format: uint32
                workId:
                    type: integer
                    description: 工号
                    format: uint32
                orgId:
                    type: integer
                    description: 部门ID
                    format: uint32
                positionId:
                    type: integer
                    description: 岗位ID
                    format: uint32
                creatorId:
                    type: integer
                    description: 创建者ID
                    format: uint32
                userName:
                    type: string
                nickName:
                    type: string
                realName:
                    type: string
                    description: 真实姓名
                avatar:
                    type: string
                    description: 头像
                email:
                    type: string
                    description: 邮箱
                phone:
                    type: string
                    description: 手机号
                gender:
                    type: string
                    description: 性别
                address:
                    type: string
                    description: 住址
                description:
                    type: string
                lastLoginTime:
                    type: string
                    description: 最后登录时间
                lastLoginIp:
                    type: string
                    description: 最后登录IP
                status:
                    enum:
                        - ON
                        - OFF
                    type: string
                    default: ON
                    description: 用户状态
                authority:
                    enum:
                        - SYS_ADMIN
                        - CUSTOMER_USER
                        - GUEST_USER
                        - REFRESH_TOKEN
                    type: string
                    default: CUSTOMER_USER
                    description: 权限
                    format: enum
                createTime:
                    type: string
                updateTime:
                    type: string
                deleteTime:
                    type: string
            description: 用户
...
    securitySchemes:
        OAuth2PasswordBearer:
            type: oauth2
            flows:
                password:
                    tokenUrl: /admin/v1/login
                    refreshUrl: /admin/v1/refresh_token
                    scopes: {}
...